Agregado manager para validar autenticacion
